Switch mob HP from int to long for large values
Updated all mob HP properties from int to long to support higher HP values for bosses and high-HP mobs. Modified Mob model, session models, and ViewModels accordingly. Added ReadVarInt64 to BinaryExtensions and updated packet processors to handle 64-bit HP. Adjusted EntityTracker logic for long HP. Updated UI ViewModels to use long for HP. Unrelated: updated file path in FilePacketCaptureDevice.

+++ AionDpsMeter.Core/Models/Mob.cs
@@ -8,8 +8,8 @@ namespace AionDpsMeter.Core.Models
         private const int BossHpThreshold = 100_000_000;
 
         public int MobCode { get; set; }
-        public int HpTotal { get; set; }
-        public int HpCurrent { get; set; }
+        public long HpTotal { get; set; }
+        public long HpCurrent { get; set; }
         //public new string Name => GetMobName();
         //public bool IsBoss => CanBeBoss();
